core: Implement libusb backend
This makes it so it's possible to implement various backend other than the serial one.
Currently, the libusb backend got implemented for compatibility with Windows (which has issues with drivers).

This also ensured compatibility for more backends in the future, and makes it so that it will be possible to implement Kamakiri as well when the time comes.